TASK: move user menu to left drawer
